pyproject.toml
[project]
name = "dental-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.1",
    "faker>=37.12.0",
    "fastapi[standard]>=0.121.0",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "nanoid>=2.0.0",
    "pydantic-ai>=1.12.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-socketio[asgi]>=5.14.3",
    "ruff>=0.14.4",
    "uvicorn>=0.38.0",
]
